Set ignore_basepython_conflict to true
Adding a basepython setting on tox makes it complains about a basepython
mismatch when the installed python version and the environment aren't
the same. Setting ignore_basepython_conflict avoids this message and
enforces the correct python version to be used.

This also needs a bump in tox version, which also comes with a feature
which lets us remove the default install setting.

See also [1]

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-May/014809.html